feat: Add module MyEnumerable
