+++ my_list_spec.rb
@@ -1,8 +1,7 @@
 require_relative 'my_list'
 
 describe MyList do
-    it 'should be a MyList' do
-        expect(MyList.new).to be_a(MyList)
-    end
-    end
-
+  it 'should be a MyList' do
+    expect(MyList.new).to be_a(MyList)
+  end
+end
